+++ lib/app/features/chat/messages/views/components/message_reaction_dialog/message_reaction_dialog.dart
@@ -1,0 +1,107 @@
+import 'dart:math';
+import 'dart:ui' as ui;
+
+import 'package:flutter/material.dart';
+import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:ion/app/components/screen_offset/screen_side_offset.dart';
+import 'package:ion/app/extensions/extensions.dart';
+import 'package:ion/app/features/chat/messages/views/components/message_reaction_dialog/components/message_reaction_context_menu.dart';
+import 'package:ion/app/features/chat/messages/views/components/message_reaction_dialog/components/message_reaction_emoji_bar.dart';
+import 'package:ion/app/services/media_service/media_service.dart';
+
+class MessageReactionDialog extends HookConsumerWidget {
+  const MessageReactionDialog({
+    required this.messagItemKey,
+    required this.isMe,
+    super.key,
+  });
+  final bool isMe;
+
+  /// The key of the message item to capture the image from widget tree
+  final GlobalKey messagItemKey;
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final capturedImage =
+        useMemoized(() async => ref.read(mediaServiceProvider).captureWidgetAsImage(messagItemKey));
+
+    return FutureBuilder(
+      future: capturedImage,
+      builder: (context, snapshot) {
+        if (!snapshot.hasData) {
+          return const SizedBox.shrink();
+        }
+
+        final (:imageBytes, :position, :size) = snapshot.data!;
+
+        /// The available height for the message content in the dialog
+        final availableHeight = MediaQuery.sizeOf(context).height -
+            MessageReactionEmojiBar.height -
+            MessageReactionContextMenu.height -
+            MediaQuery.paddingOf(context).bottom -
+            MediaQuery.paddingOf(context).top;
+
+        /// Determine the height of the message content in the dialog to fit the available height
+        final contentHeight = min(size.height, availableHeight);
+
+        /// Whether the message content is larger than the available height
+        final isHugeComponent = size.height > availableHeight;
+
+        /// The y-coordinate of the bottom of the message content
+        final bottomdY = position.dy + size.height;
+
+        /// The overflow size of the message content in the dialog
+        final overflowBottomSize = MediaQuery.sizeOf(context).height -
+            // bottomdY -
+            (position.dy > 0 ? (isHugeComponent ? 0 : bottomdY) : bottomdY) -
+            MessageReactionContextMenu.height -
+            MediaQuery.paddingOf(context).bottom;
+
+        /// The y-coordinate of the top of the message content in the dialog
+        final topY = isHugeComponent
+            ? null
+            : overflowBottomSize < 0
+                ? null
+                : position.dy - MessageReactionEmojiBar.height;
+
+        return Stack(
+          children: [
+            GestureDetector(
+              onTap: () => Navigator.of(context).pop(),
+              behavior: HitTestBehavior.opaque,
+              child: BackdropFilter(
+                filter: ui.ImageFilter.blur(sigmaX: 10, sigmaY: 10),
+                child: Container(color: context.theme.appColors.primaryText.withOpacity(0.5)),
+              ),
+            ),
+            Positioned(
+              left: isMe ? null : ScreenSideOffset.defaultSmallMargin,
+              right: isMe ? ScreenSideOffset.defaultSmallMargin : null,
+              // top: overflowBottomSize < 0 ? null : position.dy - emojiSectionHeight
+              top: topY,
+              bottom: overflowBottomSize < 0
+                  ? MediaQuery.paddingOf(context).bottom
+                  : MediaQuery.paddingOf(context).bottom,
+              // top: position.dy,
+              child: Column(
+                crossAxisAlignment: isMe ? CrossAxisAlignment.end : CrossAxisAlignment.start,
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  MessageReactionEmojiBar(isMe: isMe),
+                  Image.memory(
+                    height: contentHeight,
+                    fit: BoxFit.fitHeight,
+                    imageBytes,
+                  ),
+                  //mock context menu
+                  const MessageReactionContextMenu(),
+                ],
+              ),
+            ),
+          ],
+        );
+      },
+    );
+  }
+}
